Change license from AGPLv3 to GPLv3
The AGPL makes sense when there is a network interaction with the code (e.g.
a SaaS must share its source code to its users requesting it).

For a browser extension, it does not make sense as there is no network
interaction. The extension is installed on the user's browser.